add headers to fieldmap.base, split fieldmap.utils
